website: update docusaurus
Currently, we cannot deploy the website because trying to run Docusaurus
on a version of Node newer than v16 causes SSL errors. This commit
updates docusaurus and adds a newly required slug field to the
front-matter of the introduction page.

Node > 16 is required because older base images in Netlify have been
deprecated and installing an old version of node is very slow.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>