pkg/mesh: avoid NAT-ing packets to service CIDRs
Currently, packets to service CIDRs may be masqueraded because they are
IP addresses that Kilo does not know about and therefore is not sure if
they know about some Kilo IPs, e.g. Peer IPs. This is not terrible but
it is annoying and can prevent some advanced use-cases, see #330. This
commit adds an optional flag to the `kg` binary that can be given
multiple times to specify the service CIDRs of the cluster so that Kilo
does not masquerade packets to them.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>